Added support for LocalStorage backing for persistent data
Since `localStorage` and `sessionStorage` both provide the same API, we
can reuse our utility class to chose either of them.
