+++ rollup.config.js
@@ -42,7 +42,7 @@ export default defineConfig({
       ],
       "hyperionUtil": [
         "@hyperion/hyperion-util/src/ClientSessionID",
-        "@hyperion/hyperion-util/src/SessionPersistentData",
+        "@hyperion/hyperion-util/src/PersistentData",
         "@hyperion/hyperion-util/src/TestAndSet",
         "@hyperion/hyperion-util/src/TimedTrigger",
       ],
